refactor: naming and remove custom config option
