+++ src/lib/featureFlags/featureFlagService.ts
@@ -1,19 +1,24 @@
-import {
-  FeatureFlagConfig,
-  FeatureFlagProvider,
-} from "./featureFlagProviderShared"
+export type FeatureFlag = string
+export type FeatureFlags = Array<string>
+
+export type FeatureFlagContext = Record<string, any>
+
+export interface FeatureFlagProvider {
+  init: () => void
+  getFeatures: () => FeatureFlags
+  enabled: (name: string, context?: FeatureFlagContext) => boolean
+}
 
 const registeredFeatureFlagProviders = new Map<symbol, any>()
 
-export async function createFeatureFlagProvider<T extends FeatureFlagConfig>(
-  type: symbol,
-  config?: T
+export async function createFeatureFlagService(
+  type: symbol
 ): Promise<FeatureFlagProvider> {
   if (!registeredFeatureFlagProviders.has(type)) {
     throw new Error(`The type provider ${type.toString()} is not registered.`)
   }
 
-  // Check if provider has been registered with registerFeatureFlagProvider()
+  // Check if provider has been registered with registerFeatureFlagService()
   const featureFlagClass = registeredFeatureFlagProviders.get(type)
   if (typeof featureFlagClass !== undefined) {
     /*
@@ -21,9 +26,7 @@ export async function createFeatureFlagProvider<T extends FeatureFlagConfig>(
      ** configuration if provided, otherwise use default config from passed in
      ** featureFlagProviderClass.
      */
-    const featureFlagProvider = config
-      ? new featureFlagClass!(...config)
-      : new featureFlagClass!()
+    const featureFlagProvider = new featureFlagClass!()
 
     // initialize featureFlag client
     await featureFlagProvider.init()
@@ -35,7 +38,7 @@ export async function createFeatureFlagProvider<T extends FeatureFlagConfig>(
 
 type ConstructorType<T> = Function & { prototype: T }
 
-export function registerFeatureFlagProvider<T extends FeatureFlagProvider>(
+export function registerFeatureFlagService<T extends FeatureFlagProvider>(
   type: symbol,
   featureFlagClass: ConstructorType<T>
 ) {
