+++ src/lib/featureFlags/unleashProvider.ts
@@ -5,46 +5,10 @@ import {
   UNLEASH_INSTANCE_ID,
   UNLEASH_SERVER_KEY,
 } from "../../config"
-import {
-  FeatureFlagConfig,
-  FeatureFlagProvider,
-} from "./featureFlagProviderShared"
+import { FeatureFlagProvider } from "./featureFlagService"
 
 // Pass in as argument to registerFeatureFlagProvideder() when using unleash as feature flag service
 export const UnleashProvider = Symbol("UnleashProvider")
-
-// Override the default config set in constructor of UnleashFeatureFlagProvider class
-export class UnleashProviderConfig implements FeatureFlagConfig {
-  constructor(
-    private _url: string,
-    private _appName: string,
-    private _instanceId: string,
-    private _serverKey: string
-  ) {}
-
-  [Symbol.iterator] = function* (): Iterator<string> {
-    yield this.url
-    yield this.appName
-    yield this.instanceId
-    yield this.serverKey
-  }
-
-  get url() {
-    return this._url
-  }
-
-  get appname() {
-    return this._appName
-  }
-
-  get instanceId() {
-    return this._instanceId
-  }
-
-  get serverKey() {
-    return this._serverKey
-  }
-}
 
 // Class to instantiate Unleash client and set useful helper methods.
 export class UnleashFeatureFlagProvider implements FeatureFlagProvider {
