+++ src/v2/System/Router/index.ts
@@ -2,7 +2,7 @@ import { RelaySSREnvironment } from "v2/System/Relay/createRelaySSREnvironment"
 import { FarceCreateRouterArgs } from "found"
 import { SystemContextProps } from "../SystemContext"
 import { AppRouteConfig } from "./Route"
-import { FeatureFlags } from "lib/featureFlags/featureFlagProviderShared"
+import { FeatureFlags } from "lib/featureFlags/featureFlagService"
 
 export { Link } from "found"
 export { Boot } from "./Boot"
